requirements.txt
asgiref==3.8.1
Babel==2.15.0
beautifulsoup4==4.12.3
boto3==1.37.8
botocore==1.37.8
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==44.0.2
currency-codes==23.6.4
defusedxml==0.7.1
distlib==0.3.8
Django>4.2.16
django-allauth==65.4.1
django-autoslug==1.9.9
django-cities-light==3.10.1
django-cors-headers==4.7.0
django-countries==7.6.1
django-environ==0.12.0
django-extensions==3.2.3
django-js-asset==2.2.0
django-money==3.5.2
django-mptt==0.16.0
django-restframework==0.0.1
django-ses==4.4.0
django-smtp-ssl==1.0
django-storages==1.14.5
django-tinymce==4.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-yasg==1.21.9
filelock==3.15.1
gunicorn==23.0.0
idna==3.7
inflection==0.5.1
iso4217==1.11.20220401
jmespath==1.0.1
oauth2_provider==0.0
oauthlib==3.2.2
packaging==24.2
pillow==10.3.0
platformdirs==4.2.2
progressbar2==4.4.2
psycopg2-binary==2.9.9
py-moneyed==3.0
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-util==1.2.1
python-utils==3.8.2
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.2
requests==2.31.0
requests-oauthlib==2.0.0
s3transfer==0.11.4
setuptools==70.0.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
soupsieve==2.5
South==1.0.2
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2024.1
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.26.2
wheel==0.45.1
whitenoise==6.9.0
redis[hiredis]
django-redis
celery
django-filter
django-schema-graph
pydot==3.0.4
graphviz
weasyprint
django-sync-model
django-synchro
